3.Criando agentes com o ADK/very_simple_agent/pyproject.toml
[project]
name = "very-simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
    "google-genai>=1.65.0",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
]
